Fixed Filtering Menu Overlap moved dropdown menu component into seperate file for improved readability Added function to reference when clicks are made outside of the menu to signal menu to close, removing overlapping issue and cleaning up the page for better viewing. Issue #42
