+++ backend/package.json
@@ -14,6 +14,7 @@
     "dotenv": "^16.4.5",
     "express": "^4.21.1",
     "mongodb": "^6.9.0",
-    "mongoose": "^8.7.0"
+    "mongoose": "^8.7.0",
+    "nodemon": "^3.1.7"
   }
 }
