+++ package.json
@@ -26,6 +26,7 @@
     "@react-google-maps/api": "^2.20.3",
     "chromedriver": "^130.0.1",
     "mongodb": "^6.10.0",
+    "nodemon": "^3.1.7",
     "selenium-webdriver": "^4.25.0"
   },
   "devDependencies": {
